PRC-206 : patch contact -refactor to move patching logic into own unit
