+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/service/ContactService.kt
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.entity.ContactAddressPhoneEntity
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.entity.ContactEntity
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.mapping.patch.mapToResponse
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.mapping.patch.patchRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.mapping.toEntity
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.mapping.toModel
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.AddContactRelationshipRequest
@@ -30,6 +29,7 @@ import uk.gov.justice.digital.hmpps.hmppscontactsapi.repository.ContactPhoneDeta
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.repository.ContactRepository
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.repository.ContactSearchRepository
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.repository.PrisonerContactRepository
+import java.time.LocalDateTime
 import kotlin.jvm.optionals.getOrNull
 
 @Service
@@ -148,4 +148,38 @@ class ContactService(
     phoneNumbers: List<ContactPhoneNumberDetails>,
   ) = addressPhoneNumbers.filter { it.contactAddressId == contactAddressId }
     .mapNotNull { addressPhone -> phoneNumbers.find { it.contactPhoneId == addressPhone.contactPhoneId } }
+
+  fun ContactEntity.patchRequest(
+    request: PatchContactRequest,
+  ): ContactEntity {
+    val changedContact = this.copy(
+      title = this.title,
+      firstName = this.firstName,
+      lastName = this.lastName,
+      middleNames = this.middleNames,
+      dateOfBirth = this.dateOfBirth,
+      isDeceased = this.isDeceased,
+      deceasedDate = this.deceasedDate,
+      estimatedIsOverEighteen = this.estimatedIsOverEighteen,
+    ).also {
+      it.placeOfBirth = this.placeOfBirth
+      it.active = this.active
+      it.suspended = this.suspended
+      it.staffFlag = this.staffFlag
+      it.coronerNumber = this.coronerNumber
+      it.gender = this.gender
+      it.domesticStatus = this.domesticStatus
+      it.nationalityCode = this.nationalityCode
+      it.interpreterRequired = this.interpreterRequired
+      it.languageCode = resolveLanguageCode(this.languageCode, request.languageCode)
+      it.amendedBy = request.updatedBy
+      it.amendedTime = LocalDateTime.now()
+    }
+
+    return changedContact
+  }
+
+  fun resolveLanguageCode(dbLanguageCode: String?, requestLanguageCode: Patchable<String>?): String? {
+    return if (requestLanguageCode == null || (requestLanguageCode.get() != null)) requestLanguageCode?.get() else dbLanguageCode
+  }
 }
