+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/service/ContactServiceTest.kt
@@ -34,9 +34,6 @@ import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.ContactRelati
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.ContactSearchRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.CreateContactRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.EstimatedIsOverEighteen
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.patch.PatchContactRequest
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.patch.PatchContactResponse
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.patch.util.Patchable
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response.ContactSearchResultItem
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response.Language
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response.ReferenceCode
@@ -715,168 +712,5 @@ class ContactServiceTest {
       createdBy = "TEST",
       createdTime = LocalDateTime.now(),
     )
-  }
-
-  @Nested
-  inner class PatchContact {
-
-    @Test
-    fun `should patch a contact when fields are provided`() {
-      val contactId = 1L
-
-      val existingContact = getDummyContactEntity()
-
-      val patchRequest = patchContactRequest()
-
-      whenever(contactRepository.findById(contactId)).thenReturn(Optional.of(existingContact))
-      whenever(contactRepository.saveAndFlush(any())).thenAnswer { i -> i.arguments[0] }
-
-      val response = service.patch(contactId, patchRequest)
-
-      val contactCaptor = argumentCaptor<ContactEntity>()
-
-      verify(contactRepository).saveAndFlush(contactCaptor.capture())
-      verify(languageService).getLanguageByNomisCode("FR")
-
-      assertThat(contactCaptor.firstValue)
-        .usingRecursiveComparison()
-        .ignoringFields("amendedTime")
-        .isEqualTo(mapToEntity(patchRequest, existingContact))
-
-      assertThat(response)
-        .usingRecursiveComparison()
-        .ignoringFields("amendedTime")
-        .isEqualTo(mapToResponse(patchRequest, existingContact))
-    }
-
-    @Test
-    fun `should not validate language code if it is not available in the request`() {
-      val contactId = 1L
-
-      val existingContact = getDummyContactEntity()
-
-      val patchRequest = PatchContactRequest(
-        languageCode = null,
-        updatedBy = "Modifier",
-      )
-
-      whenever(contactRepository.findById(contactId)).thenReturn(Optional.of(existingContact))
-      whenever(contactRepository.saveAndFlush(any())).thenAnswer { i -> i.arguments[0] }
-
-      service.patch(contactId, patchRequest)
-
-      verify(languageService, never()).getLanguageByNomisCode(any())
-      verify(contactRepository, times(1)).saveAndFlush(any())
-    }
-
-    @Test
-    fun `should throw EntityNotFoundException if contact does not exist`() {
-      val contactId = 1L
-      val patchRequest = PatchContactRequest(
-        languageCode = Patchable.from("ENG"),
-        updatedBy = "system",
-      )
-
-      whenever(contactRepository.findById(contactId)).thenReturn(Optional.empty())
-
-      assertThrows<EntityNotFoundException> {
-        service.patch(contactId, patchRequest)
-      }
-    }
-  }
-
-  fun mapToResponse(request: PatchContactRequest, existingContact: ContactEntity): PatchContactResponse {
-    return PatchContactResponse(
-      id = 1L,
-      title = existingContact.title,
-      firstName = existingContact.firstName,
-      lastName = existingContact.lastName,
-      middleNames = existingContact.middleNames,
-      dateOfBirth = existingContact.dateOfBirth,
-      estimatedIsOverEighteen = existingContact.estimatedIsOverEighteen,
-      createdBy = "Admin",
-      createdTime = LocalDateTime.of(2024, 1, 22, 0, 0, 0),
-      placeOfBirth = existingContact.placeOfBirth,
-      active = existingContact.active,
-      suspended = existingContact.suspended,
-      staffFlag = existingContact.staffFlag,
-      deceasedFlag = existingContact.isDeceased,
-      deceasedDate = existingContact.deceasedDate,
-      coronerNumber = existingContact.coronerNumber,
-      gender = existingContact.gender,
-      domesticStatus = existingContact.domesticStatus,
-      languageCode = request.languageCode?.get(),
-      nationalityCode = existingContact.nationalityCode,
-      interpreterRequired = existingContact.interpreterRequired,
-      amendedBy = request.updatedBy,
-      amendedTime = LocalDateTime.now(),
-    )
-  }
-
-  fun getDummyContactEntity(): ContactEntity {
-    val existingContact = ContactEntity(
-      contactId = 1L,
-      title = "Mr.",
-      firstName = "John",
-      lastName = "Doe",
-      middleNames = "A",
-      dateOfBirth = LocalDate.of(1990, 1, 1),
-      isDeceased = false,
-      deceasedDate = null,
-      estimatedIsOverEighteen = EstimatedIsOverEighteen.DO_NOT_KNOW,
-      createdBy = "Admin",
-      createdTime = LocalDateTime.of(2024, 1, 22, 0, 0, 0),
-    ).also {
-      it.placeOfBirth = "London"
-      it.active = true
-      it.suspended = false
-      it.staffFlag = false
-      it.coronerNumber = null
-      it.gender = "M"
-      it.domesticStatus = "Single"
-      it.languageCode = "FRE-FRA"
-      it.nationalityCode = "GB"
-      it.interpreterRequired = false
-      it.amendedBy = "admin"
-      it.amendedTime = LocalDateTime.of(2024, 1, 22, 0, 0, 0)
-    }
-    return existingContact
-  }
-
-  fun mapToEntity(request: PatchContactRequest, existingContact: ContactEntity): ContactEntity {
-    return ContactEntity(
-      contactId = 1L,
-      title = existingContact.title,
-      firstName = existingContact.firstName,
-      lastName = existingContact.lastName,
-      middleNames = existingContact.middleNames,
-      dateOfBirth = existingContact.dateOfBirth,
-      estimatedIsOverEighteen = existingContact.estimatedIsOverEighteen,
-      isDeceased = existingContact.isDeceased,
-      deceasedDate = existingContact.deceasedDate,
-      createdBy = "Admin",
-      createdTime = LocalDateTime.of(2024, 1, 22, 0, 0, 0),
-    ).also {
-      it.placeOfBirth = existingContact.placeOfBirth
-      it.active = existingContact.active
-      it.suspended = existingContact.suspended
-      it.staffFlag = existingContact.staffFlag
-      it.coronerNumber = existingContact.coronerNumber
-      it.gender = existingContact.gender
-      it.domesticStatus = existingContact.domesticStatus
-      it.languageCode = request.languageCode?.get()
-      it.nationalityCode = existingContact.nationalityCode
-      it.interpreterRequired = existingContact.interpreterRequired
-      it.amendedBy = request.updatedBy
-      it.amendedTime = LocalDateTime.now()
-    }
-  }
-
-  private fun patchContactRequest(): PatchContactRequest {
-    val patchRequest = PatchContactRequest(
-      languageCode = Patchable.from("FR"),
-      updatedBy = "Modifier",
-    )
-    return patchRequest
   }
 }
